+++ src/index.ts
@@ -1,7 +1,3 @@
-// Frontmapper-Mapper Tasks
-export { MD_Frontmatter_MapBooleanInverseTask } from "./frontmatter-mapper-tasks/map-boolean-inverse-task";
-export { MD_Frontmatter_MapCopyTask } from "./frontmatter-mapper-tasks/map-copy-task";
-
 export {
   MD_Collection,
   type MD_Collection_Parameter_Type,
@@ -16,28 +12,33 @@ export {
   type MD_Frontmatter_Map,
   MD_Frontmatter_MapTask,
 } from "./md-frontmatter";
-
+// todo md-job
 export { MD_ObserverSubject, type MD_Observer_Interface } from "./md-observer";
+// todo md-tag
 export { MD_Template } from "./md-template";
-
 export { MD_Transformer_Factory } from "./md-transformer-factory";
-
 export {
   MD_Transformer_AbstractBase,
   type MD_CopyTask_Type,
   type MD_Transformer_Interface,
   type MD_Transformer_Parameter_Type,
-  type MD_Transformer_TemplateValues_Type,
 } from "./md-transformer";
 
-// Transformer Tasks
-export { MD_RemoveTODOS_Transformer } from "./transformer/md-remove-todos-task";
-export {
-  MD_ObsidianLink_Transformer_Base,
-  MD_ObsidianLink_Transformer,
-} from "./transformer/md-obsidian-link-task";
+//* Frontmapper-Mapper Tasks
+export { MD_Frontmatter_MapBooleanInverseTask } from "./frontmatter-mapper-tasks/map-boolean-inverse-task";
+export { MD_Frontmatter_MapCopyTask } from "./frontmatter-mapper-tasks/map-copy-task";
+
+//* Transformer Tasks
 export {
   MD_Splitter_Transformer,
   MD_Splitter_Parameter_Type,
 } from "./transformer/md-splitter-task";
 export { MD_Frontmatter_Transformer } from "./transformer/md-frontmatter-task";
+export { MD_Math_Transformer, type MD_MathTransformer_TemplateValues_Type } from "./transformer/md-math-task";
+export {
+  MD_ObsidianLink_Transformer_Base,
+  MD_ObsidianLink_Transformer,
+  type MD_LinkTransformer_TemplateValues_Type,
+} from "./transformer/md-obsidian-link-task";
+export { MD_RemoveTODOS_Transformer } from "./transformer/md-remove-todos-task";
+